refactor: 웹 내 컨트롤러 뷰 패키지 생성, 불필요 파일 삭제 및 파일 이동
