--- member-service/src/main/java/com/blockpage/memberservice/adaptor/web/InterestView.java
@@ -1,28 +1,0 @@
-package com.blockpage.memberservice.adaptor.web;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class InterestView {
-
-    private Long id;
-
-    private Long webtoonId;
-
-    private String webtoonTitle;
-
-    private String webtoonThumbnail;
-
-    private String creator;
-
-    private String illustrator;
-
-    private String genre;
-
-}
