+++ member-service/src/main/java/com/blockpage/memberservice/adaptor/web/apispec/APIResponse.java
@@ -11,7 +11,6 @@ public class APIResponse<T> {
 
     private final T data;
 
-
     public APIResponse(T data) {
         this.data = data;
     }
